app/utils/toon_converter.py
"""TOON format converter utilities."""
from typing import Any, Dict


class TOONConverter:
    """
    Convert data structures to TOON format for efficient LLM processing.

    TOON (Token-Oriented Object Notation) reduces token consumption
    by 30-60% compared to JSON, especially for tabular data.
    """

    @staticmethod
    def convert_to_toon(data: Any) -> str:
        """
        Convert data structure to TOON format.

        Args:
            data: Data to convert (dict, list, or primitive)

        Returns:
            TOON formatted string

        Note:
            TODO: Integrate python-toon library for proper conversion
            This is a placeholder implementation
        """
        # Placeholder - will be replaced with actual python-toon integration
        try:
            import toon
            return toon.dumps(data)
        except ImportError:
            # Fallback to basic format if toon not available
            return TOONConverter._basic_toon_format(data)

    @staticmethod
    def _basic_toon_format(data: Any, indent: int = 0) -> str:
        """Basic TOON-like format as fallback."""
        if isinstance(data, dict):
            lines = []
            for key, value in data.items():
                if isinstance(value, (dict, list)):
                    lines.append(f"{'  ' * indent}{key}:")
                    lines.append(TOONConverter._basic_toon_format(value, indent + 1))
                else:
                    lines.append(f"{'  ' * indent}{key}: {value}")
            return "\n".join(lines)
        elif isinstance(data, list):
            lines = []
            for item in data:
                lines.append(TOONConverter._basic_toon_format(item, indent))
            return "\n".join(lines)
        else:
            return f"{'  ' * indent}{data}"

    @staticmethod
    def convert_from_toon(toon_str: str) -> Any:
        """
        Convert TOON format back to Python data structure.

        Args:
            toon_str: TOON formatted string

        Returns:
            Parsed Python data structure

        Note:
            TODO: Integrate python-toon library for proper parsing
        """
        try:
            import toon
            return toon.loads(toon_str)
        except ImportError:
            # Basic parsing fallback
            return {"raw": toon_str}
