requirements.txt
# FastAPI and web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# LLM and Agent Framework
langgraph==0.2.59
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
openai==1.58.1

# Data Format
python-toon==0.1.4

# Evaluation and Metrics
streamlit==1.40.2
plotly==5.24.1
scikit-learn==1.6.0
editdistance==0.8.1
pandas==2.2.3

# Image and PDF processing
Pillow==11.0.0
pdf2image==1.17.0
PyPDF2==3.0.1

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-dotenv==1.0.1

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
